requirements.txt
biopython==1.76
flair==0.4.5
hmmlearn==0.2.3
keras==2.3.1
matplotlib==3.1.1
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
pattern3==3.0.0
scanpy==1.4.5.1
scipy==1.3.1
scikit-learn==0.21.3
seaborn==0.9.0
#tensorflow-gpu==2.2.1

anndata~=0.7.3
python-dateutil~=2.8.1
torch~=1.7.1